training/requirements.txt
imgaug==0.4.0
numpy==1.22.4
Pillow==9.5.0
torch==1.11.0+cpu
torchvision==0.12.0+cpu
